pyproject.toml
[project]
name = "emotron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitsandbytes>=0.46.1",
    "datasets>=4.0.0",
    "dspy-ai>=2.6.27",
    "liger-kernel>=0.6.1",
    "peft>=0.17.0",
    "python-dotenv>=1.1.1",
    "tensorboard>=2.20.0",
    "tiktoken>=0.11.0",
    "transformers>=4.55.0",
    "trl>=0.21.0",
]
